requirements.txt
absl-py==2.3.1
altair==6.0.0
astunparse==1.6.3
attrs==25.4.0
avro==1.11.3
blinker==1.9.0
boto3==1.42.21
botocore==1.42.21
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
confluent-kafka==2.6.1
cryptography==46.0.3
fastavro==1.11.1
flatbuffers==25.12.19
fsspec==2025.10.0
furl==2.1.4
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.76.0
h5py==3.15.1
hopsworks==4.2.10
hopsworks-aiomysql==0.2.1
idna==3.11
javaobj-py3==0.4.4
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.13.0
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mock==5.2.0
namex==0.1.0
narwhals==2.14.0
numba==0.63.1
numpy==1.26.4
opensearch-py==2.4.2
opt_einsum==3.4.0
optree==0.18.0
orderedmultidict==1.0.2
packaging==25.0
pandas==2.1.4
pillow==12.1.0
protobuf==4.25.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydeck==0.9.1
Pygments==2.19.2
pyhumps==1.6.1
pyjks==20.0.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rpds-py==0.30.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.29
streamlit==1.52.2
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
twofish==0.3.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
watchdog==6.0.0
Werkzeug==3.1.4
wrapt==2.0.1
